projeto_final_pos/requirements.txt
pandas
langchain
langchain-community
sentence-transformers
faiss-cpu
gradio
ollama
python-dotenv
pyyaml
pytest
pytest-cov
pydantic
uvicorn[standard]
